requirements.txt
assemblyai
langchain
langchain-groq
python-dotenv

streamlit>=1.32.0
